requirements.txt
torch
torchvision
pandas
matplotlib
ultralytics
Pillow
tensorflow